Move `logoutWithCleanup` to `ui/ClusterLogout`
